+++ examples/with-flow/.flowconfig
@@ -4,6 +4,5 @@
 [include]
 
 [libs]
-types/
 
 [options]
