+++ examples/with-flow/components/Page.js
@@ -1,0 +1,17 @@
+// @flow
+import React, { type Node } from 'react'
+import Head from 'next/head'
+
+type Props = {
+  children: Node,
+  title?: string
+}
+
+export default ({ children, title = 'This is the default title' }: Props) => (
+  <section>
+    <Head>
+      <title>{title}</title>
+    </Head>
+    {children}
+  </section>
+)
