+++ examples/with-flow/flow-typed/next.js.flow
@@ -1,4 +1,4 @@
-/* @flow */
+// @flow
 
 declare module "next" {
   declare type NextApp = {
