+++ examples/with-flow/pages/contact.js
@@ -1,9 +1,9 @@
 // @flow
-import * as React from 'react'
-import Layout from '../components/layout'
+import React from 'react'
+import Page from '../components/Page'
 
 export default () => (
-  <Layout title='Contact us'>
+  <Page title='Contact us'>
     <div>Contact</div>
-  </Layout>
+  </Page>
 )
